+++ src/tssim.cpp
@@ -287,8 +287,8 @@ Rcpp::List tssim(const bool tdxo = false,
                  const bool fixedFollowup = false,
                  const double plannedTime = 1350.0,
                  const double days = 30.0,
-                 const int n = 500,
-                 const int NSim = 1000, 
+                 const size_t n = 500,
+                 const size_t NSim = 1000, 
                  const int seed = 0) {
   
   std::vector<double> accTime = Rcpp::as<std::vector<double>>(accrualTime);
@@ -371,11 +371,11 @@ Rcpp::List tssim(const bool tdxo = false,
   boost::random::uniform_real_distribution<double> unif(0.0, 1.0);
   boost::random::exponential_distribution<double> expo(1.0);
   
-  int maxFollowup = static_cast<int>(std::ceil(plannedTime / days));
-  int K = n * maxFollowup;
+  size_t maxFollowup = static_cast<size_t>(std::ceil(plannedTime / days));
+  size_t K = n * maxFollowup;
   std::vector<DataFrameCpp> sims(NSim);
   
-  for (int iter = 0; iter < NSim; ++iter) {
+  for (size_t iter = 0; iter < NSim; ++iter) {
     std::vector<int> idx(K), trtrandx(K), tpointx(K);
     std::vector<int> bprogx(K), Lx(K), Llagx(K), Zx(K), Zlagx(K);
     std::vector<int> Ax(K), Alagx(K), Alag2x(K), eventx(K);
@@ -386,8 +386,8 @@ Rcpp::List tssim(const bool tdxo = false,
     
     double b1 = allocation1, b2 = allocation2;
     double enrollt = 0;
-    int k = 0;
-    for (int i = 1; i <= n; ++i) { // subject index (1..n)
+    size_t k = 0;
+    for (size_t i = 1; i <= n; ++i) { // subject index (1..n)
       int id = i;
       
       // generate accrual time
@@ -497,9 +497,9 @@ Rcpp::List tssim(const bool tdxo = false,
       if (C == 0.0) C = 1.0;
       double time = std::min({T, C, fu});
       
-      int J;  // J is the number of treatment cycles
+      size_t J;  // J is the number of treatment cycles
       if (time <= days*followup) {
-        J = static_cast<int>(std::ceil(time / days));
+        J = static_cast<size_t>(std::ceil(time / days));
         k = k - tpoint + J; // discard treatment cycles after censoring
         tstopx[k-1] = time; // update the ending time and event indicator
         eventx[k-1] = (T == time) ? 1 : 0;
@@ -529,7 +529,7 @@ Rcpp::List tssim(const bool tdxo = false,
       }
       
       // create subject-level survival time and death indicator
-      for (int j = k - J; j < k; j++) {
+      for (size_t j = k - J; j < k; j++) {
         timeOSx[j] = tstopx[k-1];
         diedx[j] = eventx[k-1];
         censor_timex[j] = fu;
@@ -538,18 +538,18 @@ Rcpp::List tssim(const bool tdxo = false,
       // progression and time to progression (if applicable)
       int pd = 0;
       double pd_time = NaN;
-      for (int j = k - J; j < k; j++) {
+      for (size_t j = k - J; j < k; j++) {
         if (Zx[j] == 1) { pd = 1; pd_time = tstopx[j]; break; }
       }
 
       // switching and time to switching (if applicable)
       int xo = 0;
       double xo_time = NaN;
-      for (int j = k - J; j < k; j++) {
+      for (size_t j = k - J; j < k; j++) {
         if (Ax[j] == 1) { xo = 1; xo_time = tstopx[j]; break; }
       }
       
-      for (int j = k - J; j < k; j++) {
+      for (size_t j = k - J; j < k; j++) {
         progressedx[j] = pd;
         timePDx[j] = pd_time;
         xox[j] = xo;
@@ -557,14 +557,14 @@ Rcpp::List tssim(const bool tdxo = false,
       }
       
       // shift disease progression and alternative therapy status downward
-      for (int j = k - J; j < k; j++) {
+      for (size_t j = k - J; j < k; j++) {
         Zx[j] = Zlagx[j];
         Ax[j] = Alagx[j];
         Alagx[j] = Alag2x[j];
       }
     }
     
-    std::vector<int> sub = seqcpp(0, k - 1);
+    std::vector<size_t> sub = seqcpp(0, k - 1);
     subset_in_place(idx, sub);
     subset_in_place(arrivalTimex, sub);
     subset_in_place( trtrandx, sub);
